Add a minimal setup.py
This file seems to be needed for Python 3.7.

Without this file, I can build and install the local directory
successfully when using Python 3.8.

When I use Python 3.7, I get the following error:
```
Directory '.' is not installable. File 'setup.py' not found.
```

According to https://stackoverflow.com/a/61762525, adding this minimal
`setup.py` will solve the issue.

It does solve the issue.