Change project description
When I switched to Windows I discovered that my Lenovo Vantage
doesn't support turning battery conservation mode on and off
anymore.

I'm not sure if it's just a bug related to my specific Windows 11
configuration, or if it's a problem for other IdeaPad users too.

I decided to make this tool compatible with Windows as I already
know the two RWEverything commands that I need to run to control
battery conservation mode in Windows. I just need to integrate
that into this Python script.

The original goals of this project (systemd and GNOME 3 integrations)
will be kept aside as I may decide to switch back to Linux in the
future.